BUG: Freelancers ratings don't update
1) After calling rate freelancer using ServerInterface the freelancers rating still return n/a.
2) The response from the server accurately gives the new rating of the freelancer.
3) The freelancers rating updates when you call freelancer.setRating(double) in the test. Line 227 is there to show that.

+++ code/Hyre_Defyer_Project_SE2/src/edu/westga/cs3211/hyre_defyer_project/server/ServerInterface.java
@@ -9,6 +9,7 @@ import org.json.JSONObject;
 import edu.westga.cs3211.hyre_defyer_project.model.Freelancer;
 import edu.westga.cs3211.hyre_defyer_project.model.FreelancerRoster;
 import edu.westga.cs3211.hyre_defyer_project.model.Message;
+import edu.westga.cs3211.hyre_defyer_project.model.RosterHelper;
 import edu.westga.cs3211.hyre_defyer_project.model.User;
 
 /**
@@ -391,13 +392,13 @@ public class ServerInterface {
 		String response = ServerCommunicator.sendRequestToServer(request);
 		JSONObject responseJSON = new JSONObject(response);
 		if (responseJSON.getString(Constants.SUCCESS_CODE).equals(Constants.REP_SUCCESS)) {
-			int rateValue = responseJSON.getInt(Constants.REP_RATING);
+			double rateValue = responseJSON.getDouble(Constants.REP_RATING);
 			for (Freelancer curr : ServerInterface.getFreelancers().getAllFreelancers()) {
 				if (curr.getUserName().equals(freelancer)) {
 					curr.setRating(rateValue);
 				}
 			}
-			return String.valueOf(responseJSON.getInt(Constants.REP_RATING));
+			return String.valueOf(responseJSON.getDouble(Constants.REP_RATING));
 
 		}
 		return "Rating failed";
